# contracts/care-coordination-platform-openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Patient Engagement Platform API
  version: 1.0.0
  description: API-first contract examples for the care coordination platform.
servers:
  - url: https://api.healthcare.example.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NotificationRequest:
      type: object
      required: [recipientId, channel, templateId, message]
      properties:
        recipientId: { type: string }
        channel:
          type: string
          enum: [SMS, EMAIL, PUSH]
        templateId: { type: string }
        message: { type: string }
      example: { recipientId: "pat-1001", channel: "SMS", templateId: "appt-reminder-v1", message: "Your appointment is tomorrow at 09:30." }
    ServiceBusMessageRequest:
      type: object
      required: [channel, eventName, payload, messageType]
      properties:
        channel: { type: string }
        eventName: { type: string }
        payload: { type: string }
        messageType: { type: string }
      example: { channel: "care-events", eventName: "AppointmentBooked", payload: "{\"appointmentId\":\"apt-2001\"}", messageType: "DOMAIN_EVENT" }
...
    MedicalRecordRequest:
      type: object
      required: [patientId, fhirResourceType, resourceReference, summary]
      properties:
        patientId: { type: string }
        fhirResourceType: { type: string }
        resourceReference: { type: string }
        summary: { type: string }
      example: { patientId: "pat-1001", fhirResourceType: "Observation", resourceReference: "Observation/obs-332", summary: "HbA1c measured at 7.4" }
    MedicalRecordUpdateRequest:
      type: object
      required: [fhirResourceType, resourceReference, summary, expectedVersion]
      properties:
        fhirResourceType: { type: string }
        resourceReference: { type: string }
        summary: { type: string }
        expectedVersion: { type: integer, minimum: 1 }
      example:
        fhirResourceType: Condition
        resourceReference: Condition/cond-332
        summary: Follow-up diagnosis updated after consultation
        expectedVersion: 1
...
    NotificationResponse:
      type: object
      required: [id, status, recipientId, channel, templateId, message, deliveryAttempts]
      properties:
        id: { type: string }
        status:
          type: string
          enum: [QUEUED, DELIVERED, FAILED]
        recipientId: { type: string }
        channel:
          type: string
          enum: [SMS, EMAIL, PUSH]
        templateId: { type: string }
        message: { type: string }
        deliveryAttempts: { type: integer, minimum: 0 }
        lastError: { type: string }
      example: { id: "not-501", status: "DELIVERED", recipientId: "pat-1001", channel: "SMS", templateId: "appt-reminder-v1", message: "Your appointment is tomorrow at 09:30.", deliveryAttempts: 1, lastError: null }
    ServiceBusMessageResponse:
      type: object
      required: [id, status, channel, eventName, payload, messageType, recordedAt, integrityHash]
      properties:
        id: { type: string }
        status:
          type: string
          enum: [QUEUED, ANOMALY_DETECTED]
        channel: { type: string }
        eventName: { type: string }
        payload: { type: string }
        messageType: { type: string }
        recordedAt: { type: string, format: date-time }
        integrityHash: { type: string }
        anomalyReason: { type: string }
      example: { id: "msg-601", status: "QUEUED", channel: "care-events", eventName: "AppointmentBooked", payload: "{\"appointmentId\":\"apt-2001\"}", messageType: "DOMAIN_EVENT", recordedAt: "2026-06-22T12:59:00Z", integrityHash: "abc123hash", anomalyReason: null }
...
    MedicalRecordResponse:
      type: object
      required: [id, status, patientId, fhirResourceType, resourceReference, summary, version]
      properties:
        id: { type: string }
        status: { type: string }
        patientId: { type: string }
        fhirResourceType: { type: string }
        resourceReference: { type: string }
        summary: { type: string }
        version: { type: integer, minimum: 1 }
      example: { id: "mr-901", status: "SYNCED", patientId: "pat-1001", fhirResourceType: "Observation", resourceReference: "Observation/obs-332", summary: "HbA1c measured at 7.4", version: 1 }
    DeviceIngestionResponse:
      type: object
      required: [id, status, deviceId, protocol, payload, receivedAt]
      properties:
        id: { type: string }
        status: { type: string }
        deviceId: { type: string }
        protocol: { type: string }
        payload: { type: string }
        receivedAt: { type: string }
      example: { id: "ing-1001", status: "INGESTED", deviceId: "dev-22", protocol: "MQTT", payload: "{\"spo2\":97}", receivedAt: "2026-05-31T10:07:00Z" }
...
    NotificationEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/NotificationResponse' }
    NotificationListEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/NotificationResponse' }
    ServiceBusMessageEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/ServiceBusMessageResponse' }
    ServiceBusMessageListEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/ServiceBusMessageResponse' }
...
    MedicalRecordEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/MedicalRecordResponse' }
    MedicalRecordListEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/MedicalRecordResponse' }
    DeviceIngestionEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/DeviceIngestionResponse' }
...
    ApiError:
      type: object
      required: [correlationId, code, message, timestamp]
      properties:
        correlationId: { type: string }
        code: { type: string }
        message: { type: string }
        timestamp: { type: string, format: date-time }
      example: { correlationId: "corr-123", code: "RESOURCE_NOT_FOUND", message: "Patient record not found: pat-404", timestamp: "2026-05-31T10:10:00Z" }
  responses:
    BadRequestResponse:
      description: Validation or request error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "VALIDATION_ERROR", message: "Validation failed", timestamp: "2026-05-31T10:11:00Z" }
    NotFoundResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "RESOURCE_NOT_FOUND", message: "Resource not found", timestamp: "2026-05-31T10:11:20Z" }
    ConflictResponse:
      description: Business rule conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "SLOT_ALREADY_BOOKED", message: "Slot is already booked", timestamp: "2026-05-31T10:11:30Z" }
    InternalServerErrorResponse:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "INTERNAL_ERROR", message: "Unexpected server error", timestamp: "2026-05-31T10:11:40Z" }
security:
  - bearerAuth: []
paths:
...
  /notifications:
    post:
      summary: Dispatch notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
            examples:
              dispatchNotification:
                value:
                  recipientId: pat-1001
                  channel: SMS
                  templateId: appt-reminder-v1
                  message: Your appointment is tomorrow at 09:30.
      responses:
        '201':
          description: Notification queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationEnvelope' }
              examples:
                created:
                  value:
                    correlationId: corr-123
                    data:
                      id: not-501
                      status: DELIVERED
                      recipientId: pat-1001
                      channel: SMS
                      templateId: appt-reminder-v1
                      message: Your appointment is tomorrow at 09:30.
                      deliveryAttempts: 1
                      lastError:
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '503': { $ref: '#/components/responses/InternalServerErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    get:
      summary: List notifications
      responses:
        '200':
          description: Notification collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationListEnvelope' }
              examples:
                list:
                  value:
                    correlationId: corr-123
                    data:
                      - id: not-501
                        status: DELIVERED
                        recipientId: pat-1001
                        channel: SMS
                        templateId: appt-reminder-v1
                        message: Your appointment is tomorrow at 09:30.
                        deliveryAttempts: 1
                        lastError:
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /notifications/{id}:
    get:
      summary: Retrieve a notification
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Notification retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      id: not-501
                      status: DELIVERED
                      recipientId: pat-1001
                      channel: SMS
                      templateId: appt-reminder-v1
                      message: Your appointment is tomorrow at 09:30.
                      deliveryAttempts: 1
                      lastError:
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /servicebus/messages:
    post:
      summary: Publish standardized message envelope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceBusMessageRequest'
            examples:
              publishEnvelope:
                value:
                  channel: care-events
                  eventName: AppointmentBooked
                  payload: '{"appointmentId":"apt-2001"}'
                  messageType: DOMAIN_EVENT
      responses:
        '201':
          description: Message queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceBusMessageEnvelope' }
              examples:
                created:
                  value:
                    correlationId: corr-123
                    data:
                      id: msg-601
                      status: QUEUED
                      channel: care-events
                      eventName: AppointmentBooked
                      payload: '{"appointmentId":"apt-2001"}'
                      messageType: DOMAIN_EVENT
                      recordedAt: 2026-06-22T12:59:00Z
                      integrityHash: abc123hash
                      anomalyReason:
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '503': { $ref: '#/components/responses/InternalServerErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    get:
      summary: List service bus messages
      responses:
        '200':
          description: Service bus message collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceBusMessageListEnvelope' }
              examples:
                list:
                  value:
                    correlationId: corr-123
                    data:
                      - id: msg-601
                        status: QUEUED
                        channel: care-events
                        eventName: AppointmentBooked
                        payload: '{"appointmentId":"apt-2001"}'
                        messageType: DOMAIN_EVENT
                        recordedAt: 2026-06-22T12:59:00Z
                        integrityHash: abc123hash
                        anomalyReason:
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /servicebus/messages/{id}:
    get:
      summary: Retrieve a service bus message
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Service bus message retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceBusMessageEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      id: msg-601
                      status: QUEUED
                      channel: care-events
                      eventName: AppointmentBooked
                      payload: '{"appointmentId":"apt-2001"}'
                      messageType: DOMAIN_EVENT
                      recordedAt: 2026-06-22T12:59:00Z
                      integrityHash: abc123hash
                      anomalyReason:
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
...
  /medical-records:
    post:
      summary: Synchronize medical record snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicalRecordRequest'
            examples:
              syncMedicalRecord:
                value:
                  patientId: pat-1001
                  fhirResourceType: Observation
                  resourceReference: Observation/obs-332
                  summary: HbA1c measured at 7.4
      responses:
        '201':
          description: Medical record synchronized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MedicalRecordEnvelope' }
              examples:
                created:
                  value:
                    correlationId: corr-123
                    data:
                      id: mr-901
                      status: SYNCED
                      patientId: pat-1001
                      fhirResourceType: Observation
                      resourceReference: Observation/obs-332
                      summary: HbA1c measured at 7.4
                      version: 1
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    get:
      summary: List medical records
      responses:
        '200':
          description: Medical record collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MedicalRecordListEnvelope' }
              examples:
                list:
                  value:
                    correlationId: corr-123
                    data:
                      - id: mr-901
                        status: SYNCED
                        patientId: pat-1001
                        fhirResourceType: Observation
                        resourceReference: Observation/obs-332
                        summary: HbA1c measured at 7.4
                        version: 1
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /medical-records/{id}:
    get:
      summary: Retrieve medical record snapshot
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Medical record retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MedicalRecordEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      id: mr-901
                      status: SYNCED
                      patientId: pat-1001
                      fhirResourceType: Observation
                      resourceReference: Observation/obs-332
                      summary: HbA1c measured at 7.4
                      version: 1
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    put:
      summary: Update medical record snapshot
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicalRecordUpdateRequest'
            examples:
              updateMedicalRecord:
                value:
                  fhirResourceType: Condition
                  resourceReference: Condition/cond-332
                  summary: Follow-up diagnosis updated after consultation
                  expectedVersion: 1
      responses:
        '200':
          description: Medical record updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MedicalRecordEnvelope' }
              examples:
                updated:
                  value:
                    correlationId: corr-123
                    data:
                      id: mr-901
                      status: SYNCED
                      patientId: pat-1001
                      fhirResourceType: Condition
                      resourceReference: Condition/cond-332
                      summary: Follow-up diagnosis updated after consultation
                      version: 2
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '409': { $ref: '#/components/responses/ConflictResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
...
  /devices/events/{id}:
    get:
      summary: Retrieve ingested device event
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Device event retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceIngestionEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      id: ing-1001
                      status: INGESTED
                      deviceId: dev-22
                      protocol: MQTT
                      payload: '{"spo2":97}'
                      receivedAt: 2026-05-31T10:07:00Z
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
...