# services/svc-notification/src/main/java/com/healthcare/notification/dto/NotificationResponse.java
package com.healthcare.notification.dto;

public record NotificationResponse(
        String id,
        String status,
        String recipientId,
        String channel,
        String templateId,
        String message,
        int deliveryAttempts,
        String lastError
) {
}